requirements.txt
astroid==2.15.5
contourpy==1.0.7
cycler==0.11.0
dill==0.3.6
distlib==0.3.6
filelock==3.12.0
fonttools==4.39.4
iniconfig==2.0.0
isort==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
matplotlib==3.7.1
mccabe==0.7.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
py4j==0.10.9.7
pylint==2.17.4
pyparsing==3.0.9
pyspark==3.4.0
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
threadpoolctl==3.1.0
tomlkit==0.11.8
tzdata==2023.3
virtualenv==20.23.0
wrapt==1.15.0
